requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Limiter==1.4
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
limits==1.5.1
MarkupSafe==1.1.1
marshmallow==3.9.1
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1
